🚚 rename all files to snake_case
#305
