+++ client/src/cb_admin/components/ActivityLabel.js
