+++ client/src/cb_admin/components/DetailsTable.js
