+++ client/src/cb_admin/components/Dropzone.js
