+++ client/src/cb_admin/components/QrBox.js
