+++ client/src/cb_admin/components/TranslucentTable.js
