+++ client/src/cb_admin/pages/Activities.js
