+++ client/src/cb_admin/pages/ConfirmPassword.js
