+++ client/src/cb_admin/pages/__tests__/Activities.test.js
