+++ client/src/cb_admin/pages/__tests__/ConfirmPassword.test.js
