+++ client/src/cb_admin/pages/__tests__/Settings.test.js
